requirements.txt
datasets
accelerate
peft
bitsandbytes
transformers
trl
sentence-transformers
huggingface_hub
faiss-gpu==1.7.2
vllm==0.3.3
gradio
tqdm
asyncio
ipywidgets 
scikit-learn 
matplotlib 
deepspeed 
sdmetrics
zipfile36
pip-autoremove
sacrebleu
unbabel-comet
lightgbm
